+++ source/include/fs.hpp
@@ -1,4 +1,4 @@
 #include <string>
 #include <cstring>
-void wf(std::string a,uint8_t *b,size_t size);
+void wf(std::string a,uint64_t *b,size_t size);
 std::string rf(std::string a);
